added videos page
videos page with most recent videos tagged with music videos, behind the scenes  and link to live videos page. Each section displays the recent 3 videos then a button that links to pages with all videos for their tag. Also added the videos and live-show sections to the nav. There is now an include for embedding videos in a post too