updated podcast page for imports
podcast page was populated by category but when the old site podcasts were tags. Also added an if statement to use the imported podcast posts image which will be thumbnail instead of image in the front matter. Podcast section has not been switched to populate by tags and use images for previous site migration