album card design tweeks
On the music page in the release cards for the albums there is a border in the design. But I noticed that the album images also had a border. So changed the image to border-bottom since the cart already had a border on the outside. I swear this wasn't there before but maybe it was.